fix(sponsorship): move sponsor into service
