+++ src/app/modules/sponsorship/sponsorship.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {
+  NbAccordionModule,
   NbButtonModule,
   NbCardModule,
   NbDialogModule,
@@ -16,12 +17,13 @@ import {
 import { AppRoutingModule } from '../routing/app-routing.module';
 import { NgxTranslateModule } from '../translate/translate.module';
 import { ShareModule } from 'src/app/components/share.module';
-import { SponsorshipService } from './sponsorship.service';
+import { SponsorshipService } from './service/sponsorship.service';
 import { TripModule } from '../trip/trip.module';
 import { CreateSponsorshipComponent } from './components/create-sponsorship/create-sponsorship.component';
+import { SelfSponsorshipsComponent } from './components/self-sponsorships/self-sponsorships.component';
 
 @NgModule({
-  declarations: [CreateSponsorshipComponent],
+  declarations: [CreateSponsorshipComponent, SelfSponsorshipsComponent],
   imports: [
     AppRoutingModule,
     CommonModule,
@@ -31,6 +33,7 @@ import { CreateSponsorshipComponent } from './components/create-sponsorship/crea
     NbInputModule,
     NbButtonModule,
     NbCardModule,
+    NbAccordionModule,
     NbListModule,
     NbIconModule,
     NbSpinnerModule,
@@ -41,6 +44,6 @@ import { CreateSponsorshipComponent } from './components/create-sponsorship/crea
     TripModule
   ],
   providers: [NbLayoutRulerService, NbLayoutRulerService, SponsorshipService],
-  exports: [CreateSponsorshipComponent]
+  exports: [CreateSponsorshipComponent, SelfSponsorshipsComponent]
 })
 export class SponsorshipModule {}
